+++ mrb_doc/core/input/mouse.rb
@@ -1,8 +1,0 @@
-# Sets the mouse cursor.
-# @param cursor [Integer]
-# @return [nil]
-def set_mouse_cursor(cursor)
-  # mrb_set_mouse_cursor
-  # src/mruby_integration/core/input/mouse.cpp
-  nil
-end